Got paging to work by copypastring from jamesmolly tutorial, added in generic bitwise operations, isr.c: automatically send EOI to PIC interrupt and halt the machine if an unhandled processor exception occurrs, paging.c/h: fix typos and small errors, keyboard.c: no need to send EOI now, kmalloc.c: fixed wrong address return type

+++ kernel/kmalloc.c
@@ -5,7 +5,7 @@ udword start_free_mem = 0x200000;
 //...and end at 2MB mark
 udword end_free_mem = 0x300000;
 
-void *kmalloc(int size, int align, unsigned short *phys){
+void *kmalloc(int size, int align, unsigned int *phys){
    if( start_free_mem + size < end_free_mem ){
       if( align ){
         start_free_mem &= 0xFFFFF000;  //Align the allocated memory
