+++ kernel/kmalloc.h
@@ -5,4 +5,4 @@
 
 //udword start_free_mem = 0x1000;
 //udword kmalloc(udword size, int align, udword *phys_addr);
-void *kmalloc(int,int,unsigned short*);
+void *kmalloc(int,int,unsigned int*);
